feat: use model module
